+++ lib/office365/rest/contact.rb
@@ -1,13 +1,21 @@
 # frozen_string_literal: true
 
+require_relative "./concerns/base"
+
 module Office365
   module REST
     module Contact
-      def contacts
-        contacts_uri = ["/", Office365::API_VERSION, "/me/contacts"].join
-        response = Request.new(access_token, debug: debug).get(contacts_uri)
+      include Concerns::Base
 
-        response["value"].map { |v| Models::User.new(v) }
+      # params: args => { next_link: (nil / next_page_url) }
+      # response { results: [], next_link: '...' }
+      def contacts(args = {})
+        response = message_response(args: args.merge(base_uri: "/me/contacts"))
+
+        {
+          results: response["value"].map { |v| Models::Contact.new(v) },
+          next_link: response["@odata.nextLink"]
+        }
       end
     end
   end